Enable resolver resolution to use function to generate a name.
This small hook means that some of our cases where we use a different
resolver based on user input can be handled by writing a helper function
to choose a resolver by name.

In addition, I added a "null" resolver which a) helps with testing and
b) helps with cases where one of the branch paths for the above hook
wants to return null instead of resolving something.