Enable resolver pipelines.
Resolvers have accumulated several sub-functions, including `mask` functions (which
manipulate inputs) and `transform` functions (which manipulate outputs). These concepts
work well in the context of a single resolver, but fall short when reusing "core"
resolvers in multiple contexts with slight differences.

In these cases, we want the ability to "attach" custom `mask` and `transform` functions
to an existing resolver. The mechanism we use is a *pipeline* compose of a list of keys
(or key functions) that lookup a mask, resolver, or transform function in the DI container.

For example, this pipeline:

```
getResolverPipeline('extractCompanyId', 'company.retrieve', 'toList')
```

Would most likely:
 -  Apply a mask to the input `obj` value to fetch a `companyId`
 -  Invoke the company retrieve resolver to get a company
 -  Transform the company into list form

Meaning that:
 -  Company retrieval can focus on doing a single job
 -  List transformation can be reused