Pass unmasked parameters to authorize
After some experience with larger configurations, the value of `mask` is to simplify
invocations to `aggregate` (and to a lesser extent `transform`). It turns out to be
useful to have different masks for different resolvers... but to have unified
authorization implementations. So: pass the pre-masked arguments.

Not backwards compatible, but our current surface area for authorization is small.