GLOB-26190 - moving cache code into nodule-graphql
Why?
Our FE gateway is where we tested our caching code, and now that proved itself, we
want to reuse it in the rest of our gateways. `nodule-graphql` is where both our
dedup and batching service wrappers reside, so it's where the new code will live as well.

What I've done:
1. Moved caching service wrapper code from our FE gateway
2. Refactored a couple of piece of that code (using default validators as functions, separating logging, etc)
3. Removed references to globality from tests and mocks (using Rick & Morty terms)
4. Wrote a new README item in the services README, explaining the way to turn caching on