Middleware resolvers
There are action that we may want to perform for all (or almost all) resolvers in an app.
Currently the only way to do that is to add this action to every resolver. Middleware resolvers allow us to instead centrally define both those actions and the resolvers to which they shouldn't apply.
- Create the MiddlewareResolver class and its interface
- Execute middleware resolvers in any resolver pipeline, before the actual resolver logic.